Sprite direction and counter
Added feature where the sprite now faces the direction of the mouse cursor, or on controller, the direction of movement, whilst being able to manually switch the facing direction with the right stick.

Added a very basic placeholder for the counter, which will be better explored once damage dealing features are added.

Added new mappings into the project, as well as tweaked some other QoL settings.