Added Mutated Bat Animations
Implemented all the animations, timing, positioning, etc. for the mutated bat. The current animations will soon be switched for ones without wind, however nothing else will change.

Additionally, made the end of the swoop a bit cleaner, with the bat continuing forwards with a velocity afterwards, instead of just stopping at the end.