Fully Finished Wolf AI
Wolf AI fully finished. This includes the patrol (within a given patrol area), an aggressive state, an occasional jump bit attack, and properly configurated movement.

Did some tweaking to the MutatedBat animations, and identified problems to fix in another branch.

Finally, went on a short side quest and fixed attacking whilst clicking in the inventory, by checking if the cursor was in the inventory on click.