Better camera
Added manually-coded smoothing to the camera to give more control and a better user experience compared to the default settings.

Additionally, zoomed out the player camera to better fit the type of game.

Finally, made the PlayerCamera a child of the Player scene, instead of needing to add it whenever we instantiate the player.