Finished Text System
Added NPC and Readable text systems. For both, pressing the Input key will open the dialogue box.

NPCs work using a JSON that includes dialogue checkpoints, nodes, actions, player dialogue options, etc. These are set on a per NPC basis, as is each NPC's script.

Readables work much simpler, and text is built-in to the script of each readable. This may however change to be more robust, as with NPCs.

This is all managed in the Autoload DialogueManager script.